refactor: extract internal/watch package
Move the per-root fsnotify watcher manager and watch loop out of the
taskfileserver package and into a new internal/watch package. The
watch package depends only on stdlib + fsnotify + slog plus a small
StateProvider interface satisfied by *Server, so it has no knowledge
of the orchestrator's concrete type.

What moved:
- watcherManager -> watch.Manager, with exported Apply/Reconcile/
  Shutdown/Active/IsShuttingDown methods
- watchRootTaskfiles -> watch.Watch (StateProvider-driven loop)
- watcher_manager_test.go -> internal/watch/manager_test.go (whole file)
- syncWatcherDirs (helper, unexported in watch)

What stayed:
- Integration-style watch_test.go remains in taskfileserver because it
  asserts Server-level reload behaviour
- A small internal/taskfileserver/watch.go retains Server.Shutdown,
  Server.restartWatchers, and Server.RootWatchState (which now
  satisfies watch.StateProvider)

Renames to satisfy the StateProvider interface:
- Server.reloadRoot -> Server.ReloadRoot
- Server.rootWatchState -> Server.RootWatchState

Server.New now wires the manager via watch.New(s, s.log); the
LoggerFunc indirection means SetLogger continues to work without
restarting watchers. Public surface (New, SetLogger, SetToolRegistry,
HandleInitialized, HandleRootsChanged, Shutdown) is unchanged and
main.go is untouched.

Preserved log events watcher.reload_failed, watcher.fs_error, and
watcher.failed with their original fields. No behaviour change; task
ci is green.

Refs #80
